Replace all systemd-boot boot modes with a single one
To simplify boot modes, deprecate `uefi-x64.systemd-boot.eltorito`,
`uefi-x64.systemd-boot.esp`, `uefi-ia32.systemd-boot.eltorito` and
`uefi-ia32.systemd-boot.eltorito`, and add a combined
`uefi.systemd-boot` instead.

The ability to create ISOs without El Torito UEFI boot entries is lost,
but if such a thing were to be desired, then there would be no reason to
use the ISO format in the first place. Instead of something like a more
simple raw disk image (to be written with `dd`) or, alternatively, a
tarball (that would need to be manually extracted) could be used instead.
Since virtual machine software expect El Torito ISOs, this is not even
worth considering.

The simplified `uefi.systemd-boot` boot mode makes it easier to strive
for an automatic `$arch` to UEFI architecture matching in the future.